Splited FVC constraints into another file, and now tableViewModel is also the datasource
